refactor: extract _normalize_coords to LocateOperations
Move coordinate validation/normalization method (~75 lines) from main.py
to locate.py.
